Started the binary/decimal converter.
Only worked on the input validation part of the program.
The actual converter is yet to be started.

For input validation the last thing that has to be done is to understand how variable shadow eachother between loops to fix the code which in theory is already done but has some implementation problems.